backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
pydantic==2.9.2
cryptography==43.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
pyyaml==6.0.2
httpx==0.27.2
